requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-extra-fields==1.2.4
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
drf-yasg==1.16.1
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
simplejwt==2.0.1
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.5
zipp==0.6.0
